features branch new message
Add message indicating first commit to console output

Updated the `Main` method in `Program.cs` to include an additional `Console.WriteLine` statement. This new line outputs the message "This is my first commit!" to the console, providing additional feedback to the user when the application runs.

Add message indicating first commit to console output

Updated the `Main` method in `Program.cs` to include an additional `Console.WriteLine` statement. This new line outputs the message "This is my first commit!" to the console, providing additional feedback to the user when the application runs.

Add message indicating first commit to console output

Updated the `Main` method in `Program.cs` to include an additional `Console.WriteLine` statement. This new line outputs the message "This is my first commit!" to the console, providing additional feedback to the user when the application runs.